+++ src/main.h
@@ -19,7 +19,7 @@
 #pragma once
 
 #ifdef BUILD_REL
-    #define MOD_VERSION "Thinker Mod v5.1 - The Will to Power mod v381"
+    #define MOD_VERSION "Thinker Mod v5.1 - The Will to Power mod v382"
 #else
     #define MOD_VERSION "Thinker Mod develop build"
 #endif
@@ -443,6 +443,7 @@ struct Config {
     bool display_numeric_mood = false;
     int manifold_harmonics_min_planet_rating = 0;
     bool fungus_terrain_bonus = false;
+    int promotion_probability_denominator = 1;
     // AI configurations
     bool ai_useWTPAlgorithms;
     bool wtp_enabled_factions[MaxPlayerNum];